2023 AoC06 complete solution
There were many ways to solve this. I did 2 different ways.

The first one is iterative so not super efficienct. I know the first
derivative of a function is the max point (if you set the derivate = to
0 and solve) so I did that and rounded to give me the nearest integer as
  my midpoint and then I walked that until the distance I was finding
  was less than the record holders. This works -- not efficiency really
  could've done a binary search walk and it would've been better.

Second solution was to use the quadratic equation on the bigger data.
A line is Y = MX + B this function mapping time to distance happened to
be quadratic e.g. Y = AX^2 + BX + C.

Specifically, it was the function F(X) = X(C-X) or -X^2 + XC

Setting the line equation equal to the quadratic equation and solving
eventually yields:

CrossingPoints = (-time +/- SQRT( time^2 - 4 * dist))/-2

After that I built a range but had to remember that python's range
function is exlusive so I used ceilings on both roots to find the full
range.